Update the resulted back

+++ Application/BlogPosts/Delete.cs
@@ -1,4 +1,5 @@
-﻿using MediatR;
+﻿using Application.Core;
+using MediatR;
 using Persistence;
 using System;
 using System.Threading;
@@ -8,12 +9,12 @@ namespace Application.BlogPosts
 {
     public class Delete
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
 
@@ -21,14 +22,20 @@ namespace Application.BlogPosts
             {
                 _context = context;
             }
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var blogPost = await _context.BlogPosts.FindAsync(request.Id);
 
-                _context.BlogPosts.Remove(blogPost);
-                await _context.SaveChangesAsync();
+                if (blogPost == null)
+                    return null;
 
-                return Unit.Value;
+                _context.BlogPosts.Remove(blogPost);
+                var result = await _context.SaveChangesAsync() > 0;
+
+                if (!result)
+                    return Result<Unit>.Failure("Failed to delete the blog post");
+
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
